refactor: move connector constants to @neoboard/connection, add dep to component
- Create connection/src/connector-types.ts as canonical source of truth
- Export from connection/src/index.ts
- Add @neoboard/connection as dep in component/package.json
- app/src/lib/connector-types.ts now re-exports from @neoboard/connection

Both app and component can now import connector constants from the
connection package, eliminating string coupling.
