+++ component/package.json
@@ -34,6 +34,7 @@
     "@hookform/resolvers": "^5.2.2",
     "@neo4j-cypher/language-support": "^2.0.0-next.30",
     "@neo4j-nvl/react": "^1.1.0",
+    "@neoboard/connection": "file:../connection",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-avatar": "^1.1.11",
